Beta Version 3.5
Build: 25125 |
https://rink.hockeyapp.net/apps/0172d48cceec171249a8d850fb16276b

+++ Headers/MSDistributeAction.h
@@ -11,8 +11,8 @@
 }
 
 - (void)clearCachesIfNecessary:(id)arg1;
-- (void)distributeLayers:(id)arg1 withSpace:(long long)arg2 forAxis:(unsigned long long)arg3;
-- (long long)calculateRequiredSpaceBetweenLayers:(id)arg1 forAxis:(unsigned long long)arg2;
+- (void)distributeLayers:(id)arg1 withSpace:(double)arg2 forAxis:(unsigned long long)arg3;
+- (double)calculateRequiredSpaceBetweenLayers:(id)arg1 forAxis:(unsigned long long)arg2;
 - (id)layersSortedForAxis:(unsigned long long)arg1;
 - (void)distributeAlongAxis:(unsigned long long)arg1;
 - (BOOL)validate;
